More moves and renames.
