+++ src/extension.ts
@@ -35,9 +35,9 @@ import { wrapperResolver } from './resolvers/WrapperResolver';
 import { AzureAccountTreeItem } from './tree/AzureAccountTreeItem';
 import { GroupTreeItemBase } from './tree/GroupTreeItemBase';
 import { HelpTreeItem } from './tree/HelpTreeItem';
-import { ApplicationResourceBranchDataProviderManager } from './tree/v2/providers/ApplicationResourceBranchDataProviderManager';
-import { DefaultApplicationResourceBranchDataProvider } from './tree/v2/providers/DefaultApplicationResourceBranchDataProvider';
-import { registerResourceGroupsTreeV2 } from './tree/v2/registerResourceGroupsTreeV2';
+import { ApplicationResourceBranchDataProviderManager } from './tree/v2/application/ApplicationResourceBranchDataProviderManager';
+import { DefaultApplicationResourceBranchDataProvider } from './tree/v2/application/DefaultApplicationResourceBranchDataProvider';
+import { registerResourceGroupsTreeV2 } from './tree/v2/application/registerResourceGroupsTreeV2';
 import { registerWorkspaceTreeV2 } from './tree/v2/workspace/registerWorkspaceTreeV2';
 import { WorkspaceDefaultBranchDataProvider } from './tree/v2/workspace/WorkspaceDefaultBranchDataProvider';
 import { WorkspaceResourceBranchDataProviderManager } from './tree/v2/workspace/WorkspaceResourceBranchDataProviderManager';
