+++ src/tree/v2/application/ApplicationResourceBranchDataProviderManager.ts
