+++ src/tree/v2/application/DefaultApplicationResourceItem.ts
