+++ src/tree/v2/application/GroupingItem.ts
@@ -1,11 +1,11 @@
 import { TreeItemIconPath } from '@microsoft/vscode-azext-utils';
 import * as vscode from 'vscode';
-import { ApplicationResource, BranchDataProvider, ResourceModelBase } from '../../api/v2/v2AzureResourcesApi';
-import { getIconPath } from '../../utils/azureUtils';
-import { BranchDataItemFactory } from './BranchDataItem';
-import { BranchDataProviderFactory } from './providers/ApplicationResourceBranchDataProviderManager';
-import { ResourceGroupsItem } from './ResourceGroupsItem';
-import { ResourceGroupsTreeContext } from './ResourceGroupsTreeContext';
+import { ApplicationResource, BranchDataProvider, ResourceModelBase } from '../../../api/v2/v2AzureResourcesApi';
+import { getIconPath } from '../../../utils/azureUtils';
+import { BranchDataItemFactory } from '../BranchDataItem';
+import { BranchDataProviderFactory } from './ApplicationResourceBranchDataProviderManager';
+import { ResourceGroupsItem } from '../ResourceGroupsItem';
+import { ResourceGroupsTreeContext } from '../ResourceGroupsTreeContext';
 
 export class GroupingItem implements ResourceGroupsItem {
     private description: string | undefined;
