+++ src/tree/v2/application/SubscriptionItem.ts
@@ -1,10 +1,10 @@
 import * as vscode from "vscode";
-import { ApplicationResourceProviderManager } from "../../api/v2/ApplicationResourceProviderManager";
-import { ApplicationSubscription } from "../../api/v2/v2AzureResourcesApi";
-import { treeUtils } from "../../utils/treeUtils";
+import { ApplicationResourceProviderManager } from "../../../api/v2/ApplicationResourceProviderManager";
+import { ApplicationSubscription } from "../../../api/v2/v2AzureResourcesApi";
+import { treeUtils } from "../../../utils/treeUtils";
 import { ApplicationResourceGroupingManager } from "./ApplicationResourceGroupingManager";
-import { ResourceGroupsItem } from "./ResourceGroupsItem";
-import { ResourceGroupsTreeContext } from "./ResourceGroupsTreeContext";
+import { ResourceGroupsItem } from "../ResourceGroupsItem";
+import { ResourceGroupsTreeContext } from "../ResourceGroupsTreeContext";
 
 export class SubscriptionItem implements ResourceGroupsItem {
     constructor(
